+++ airflow/contrib/operators/bigquery_get_data.py
@@ -16,122 +16,14 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-"""
-This module contains a Google BigQuery data operator.
-"""
+"""This module is deprecated. Please use `airflow.gcp.operators.bigquery`."""
+
 import warnings
-from typing import Optional
 
-from airflow.contrib.hooks.bigquery_hook import BigQueryHook
-from airflow.models import BaseOperator
-from airflow.utils.decorators import apply_defaults
+# pylint: disable=unused-import
+from airflow.gcp.operators.bigquery import BigQueryGetDataOperator  # noqa
 
-
-class BigQueryGetDataOperator(BaseOperator):
-    """
-    Fetches the data from a BigQuery table (alternatively fetch data for selected columns)
-    and returns data in a python list. The number of elements in the returned list will
-    be equal to the number of rows fetched. Each element in the list will again be a list
-    where element would represent the columns values for that row.
-
-    **Example Result**: ``[['Tony', '10'], ['Mike', '20'], ['Steve', '15']]``
-
-    .. note::
-        If you pass fields to ``selected_fields`` which are in different order than the
-        order of columns already in
-        BQ table, the data will still be in the order of BQ table.
-        For example if the BQ table has 3 columns as
-        ``[A,B,C]`` and you pass 'B,A' in the ``selected_fields``
-        the data would still be of the form ``'A,B'``.
-
-    **Example**: ::
-
-        get_data = BigQueryGetDataOperator(
-            task_id='get_data_from_bq',
-            dataset_id='test_dataset',
-            table_id='Transaction_partitions',
-            max_results='100',
-            selected_fields='DATE',
-            gcp_conn_id='airflow-conn-id'
-        )
-
-    :param dataset_id: The dataset ID of the requested table. (templated)
-    :type dataset_id: str
-    :param table_id: The table ID of the requested table. (templated)
-    :type table_id: str
-    :param max_results: The maximum number of records (rows) to be fetched
-        from the table. (templated)
-    :type max_results: str
-    :param selected_fields: List of fields to return (comma-separated). If
-        unspecified, all fields are returned.
-    :type selected_fields: str
-    :param gcp_conn_id: (Optional) The connection ID used to connect to Google Cloud Platform.
-    :type gcp_conn_id: str
-    :param bigquery_conn_id: (Deprecated) The connection ID used to connect to Google Cloud Platform.
-        This parameter has been deprecated. You should pass the gcp_conn_id parameter instead.
-    :type bigquery_conn_id: str
-    :param delegate_to: The account to impersonate, if any.
-        For this to work, the service account making the request must have domain-wide
-        delegation enabled.
-    :type delegate_to: str
-    :param location: The location used for the operation.
-    :type location: str
-    """
-    template_fields = ('dataset_id', 'table_id', 'max_results')
-    ui_color = '#e4f0e8'
-
-    @apply_defaults
-    def __init__(self,
-                 dataset_id: str,
-                 table_id: str,
-                 max_results: str = '100',
-                 selected_fields: Optional[str] = None,
-                 gcp_conn_id: str = 'google_cloud_default',
-                 bigquery_conn_id: Optional[str] = None,
-                 delegate_to: Optional[str] = None,
-                 location: str = None,
-                 *args,
-                 **kwargs) -> None:
-        super().__init__(*args, **kwargs)
-
-        if bigquery_conn_id:
-            warnings.warn(
-                "The bigquery_conn_id parameter has been deprecated. You should pass "
-                "the gcp_conn_id parameter.", DeprecationWarning, stacklevel=3)
-            gcp_conn_id = bigquery_conn_id
-
-        self.dataset_id = dataset_id
-        self.table_id = table_id
-        self.max_results = max_results
-        self.selected_fields = selected_fields
-        self.gcp_conn_id = gcp_conn_id
-        self.delegate_to = delegate_to
-        self.location = location
-
-    def execute(self, context):
-        self.log.info('Fetching Data from:')
-        self.log.info('Dataset: %s ; Table: %s ; Max Results: %s',
-                      self.dataset_id, self.table_id, self.max_results)
-
-        hook = BigQueryHook(bigquery_conn_id=self.gcp_conn_id,
-                            delegate_to=self.delegate_to,
-                            location=self.location)
-
-        conn = hook.get_conn()
-        cursor = conn.cursor()
-        response = cursor.get_tabledata(dataset_id=self.dataset_id,
-                                        table_id=self.table_id,
-                                        max_results=self.max_results,
-                                        selected_fields=self.selected_fields)
-
-        self.log.info('Total Extracted rows: %s', response['totalRows'])
-        rows = response['rows']
-
-        table_data = []
-        for dict_row in rows:
-            single_row = []
-            for fields in dict_row['f']:
-                single_row.append(fields['v'])
-            table_data.append(single_row)
-
-        return table_data
+warnings.warn(
+    "This module is deprecated. Please use `airflow.gcp.operators.bigquery`.",
+    DeprecationWarning,
+)
