+++ tests/test_core_to_contrib.py
@@ -102,6 +102,10 @@ class TestMovingCoreToContrib(TestCase):
                 "airflow.contrib.hooks.gcp_sql_hook.CloudSqlHook",
             ),
             (
+                "airflow.gcp.hooks.cloud_sql.CloudSqlDatabaseHook",
+                "airflow.contrib.hooks.gcp_sql_hook.CloudSqlDatabaseHook"
+            ),
+            (
                 "airflow.gcp.hooks.tasks.CloudTasksHook",
                 "airflow.contrib.hooks.gcp_tasks_hook.CloudTasksHook",
             ),
@@ -128,6 +132,10 @@ class TestMovingCoreToContrib(TestCase):
             (
                 "airflow.gcp.hooks.dataproc.DataProcHook",
                 "airflow.contrib.hooks.gcp_dataproc_hook.DataProcHook",
+            ),
+            (
+                "airflow.gcp.hooks.bigquery.BigQueryHook",
+                "airflow.contrib.hooks.bigquery_hook.BigQueryHook",
             ),
         ]
     )
@@ -663,6 +671,58 @@ class TestMovingCoreToContrib(TestCase):
                 "airflow.contrib.operators.dataproc_operator."
                 "DataprocWorkflowTemplateInstantiateOperator",
             ),
+            (
+                "airflow.gcp.operators.bigquery.BigQueryCheckOperator",
+                "airflow.contrib.operators.bigquery_check_operator.BigQueryCheckOperator",
+            ),
+            (
+                "airflow.gcp.operators.bigquery.BigQueryIntervalCheckOperator",
+                "airflow.contrib.operators.bigquery_check_operator.BigQueryIntervalCheckOperator",
+            ),
+            (
+                "airflow.gcp.operators.bigquery.BigQueryValueCheckOperator",
+                "airflow.contrib.operators.bigquery_check_operator.BigQueryValueCheckOperator",
+            ),
+            (
+                "airflow.gcp.operators.bigquery.BigQueryGetDataOperator",
+                "airflow.contrib.operators.bigquery_get_data.BigQueryGetDataOperator",
+            ),
+            (
+                "airflow.gcp.operators.bigquery.BigQueryCreateEmptyDatasetOperator",
+                "airflow.contrib.operators.bigquery_operator.BigQueryCreateEmptyDatasetOperator",
+            ),
+            (
+                "airflow.gcp.operators.bigquery.BigQueryCreateEmptyTableOperator",
+                "airflow.contrib.operators.bigquery_operator.BigQueryCreateEmptyTableOperator",
+            ),
+            (
+                "airflow.gcp.operators.bigquery.BigQueryCreateExternalTableOperator",
+                "airflow.contrib.operators.bigquery_operator.BigQueryCreateExternalTableOperator",
+            ),
+            (
+                "airflow.gcp.operators.bigquery.BigQueryDeleteDatasetOperator",
+                "airflow.contrib.operators.bigquery_operator.BigQueryDeleteDatasetOperator",
+            ),
+            (
+                "airflow.gcp.operators.bigquery.BigQueryOperator",
+                "airflow.contrib.operators.bigquery_operator.BigQueryOperator",
+            ),
+            (
+                "airflow.gcp.operators.bigquery.BigQueryTableDeleteOperator",
+                "airflow.contrib.operators.bigquery_table_delete_operator.BigQueryTableDeleteOperator",
+            ),
+            (
+                "airflow.operators.bigquery_to_bigquery.BigQueryToBigQueryOperator",
+                "airflow.contrib.operators.bigquery_to_bigquery.BigQueryToBigQueryOperator",
+            ),
+            (
+                "airflow.operators.bigquery_to_gcs.BigQueryToCloudStorageOperator",
+                "airflow.contrib.operators.bigquery_to_gcs.BigQueryToCloudStorageOperator",
+            ),
+            (
+                "airflow.operators.bigquery_to_mysql.BigQueryToMySqlOperator",
+                "airflow.contrib.operators.bigquery_to_mysql_operator.BigQueryToMySqlOperator",
+            ),
         ]
     )
     def test_operators_paths(self, new_path: str, old_path: str):
@@ -685,6 +745,10 @@ class TestMovingCoreToContrib(TestCase):
                 "airflow.gcp.sensors.pubsub.PubSubPullSensor",
                 "airflow.contrib.sensors.pubsub_sensor.PubSubPullSensor",
             ),
+            (
+                "airflow.gcp.sensors.bigquery.BigQueryTableSensor",
+                "airflow.contrib.sensors.bigquery_sensor.BigQueryTableSensor",
+            )
         ]
     )
     def test_sensor_paths(self, new_path: str, old_path: str):
